Migrated UrlPathPrepending bot.
