+++ configuration/application.config
@@ -1,5 +1,6 @@
 {
   "bot": {
+    "root_url": "http://play.pokemonshowdown.com/",
     "disposer": {
       "generator": {
         "file_path": "/var/opt/downloadbot/log/screenshot.png"
